Move wrapper to schema dir
