+++ breadbox/breadbox/io/filestore_crud.py
@@ -5,7 +5,7 @@ from typing import Any, List, Optional, Union
 
 import pandas as pd
 
-from .data_validation import DataFrameWrapper
+from ..schemas.dataframe_wrapper import DataFrameWrapper
 from ..models.dataset import Dataset, MatrixDataset, ValueType
 from .hdf5_utils import write_hdf5_file, read_hdf5_file
