+++ breadbox/breadbox/schemas/dataframe_wrapper.py
@@ -1,0 +1,68 @@
+from typing import Protocol, List
+import pandas as pd
+import numpy as np
+import pyarrow.parquet as pq
+
+
+class DataFrameWrapper(Protocol):
+    def get_index_names(self) -> List[str]:
+        ...
+
+    def get_column_names(self) -> List[str]:
+        ...
+
+    def read_columns(self, columns: list[str]) -> pd.DataFrame:
+        ...
+
+    def is_sparse(self) -> bool:
+        ...
+
+
+class ParquetDataFrameWrapper:
+    def __init__(self, parquet_path: str):
+        self.parquet_path = parquet_path
+        self.schema = pq.read_schema(parquet_path)
+
+    def get_index_names(self) -> List[str]:
+        index_col = self.schema.names[0]
+        print("index_col", index_col)
+        index_names = pd.read_parquet(
+            self.parquet_path, columns=[index_col], engine="pyarrow"
+        )[index_col].to_list()
+        print("index_names", index_names[:10])  # Print first 10 for debugging
+        return index_names
+
+    def get_column_names(self) -> List[str]:
+        col_names = self.schema.names[1:]  # Exclude the index column
+        print("col_names", col_names[:10])  # Print first 10 for debugging
+        return col_names
+
+    def read_columns(self, columns: list[str]) -> pd.DataFrame:
+        return pd.read_parquet(self.parquet_path, columns=columns, engine="pyarrow")
+
+    def is_sparse(self) -> bool:
+        # For now, we bypass checking sparsity for Parquet files to reduce complexity.
+        return False
+
+
+class PandasDataFrameWrapper:
+    def __init__(self, df: pd.DataFrame):
+        self.df = df
+        # Get the row,col positions where df values are not null
+        rows_idx, cols_idx = np.where(df.notnull())
+        self.nonnull_indices = list(zip(rows_idx, cols_idx))
+
+    def get_index_names(self) -> List[str]:
+        return self.df.index.to_list()
+
+    def get_column_names(self) -> List[str]:
+        return self.df.columns.to_list()
+
+    def read_columns(self, columns: list[str]) -> pd.DataFrame:
+        return self.df.loc[:, columns]
+
+    def is_sparse(self) -> bool:
+        total_nulls = self.df.size - len(self.nonnull_indices)
+        # Determine whether matrix is considered sparse (~2/3 elements are null). Use chunked storage for sparse matrices for more optimal storage
+        is_sparse = total_nulls / self.df.size > 0.6
+        return is_sparse
